requirements.txt
flask>=3.0.0
PyJWT>=2.7.0
Werkzeug>=3.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
